% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdrive.R
\name{ggdrive}
\alias{ggdrive}
\title{Create Drive Chart from nflverse PBP}
\usage{
ggdrive(
  pbp,
  type = c("all", "away", "home"),
  with_score_plot = FALSE,
  home_reverse = TRUE
)
}
\arguments{
\item{pbp}{nflverse play-by-play data filtered down to one game}

\item{type}{One of "all", "away", "home". Decides what type of chart to build}

\item{with_score_plot}{Either \code{TRUE} or \code{FALSE}. Choose to add scores to the plot.}

\item{home_reverse}{Reverse direction of home team for chart \code{type} "home"}
}
\value{
A ggplot2 object
}
\description{
Create Drive Chart from nflverse PBP
}
\examples{
pbp <- nflreadr::load_pbp(2022)

g <- dplyr::filter(pbp, game_id == "2022_05_LV_KC")

ggdrive(g)
ggdrive(g, "away", with_score_plot = TRUE)
ggdrive(g, "home", with_score_plot = TRUE)
}
